requirements.txt
torch
torchmetrics
streamlit
nltk
numpy
pandas
scikit-learn
ipython==8.31.0
openai
python-dotenv